gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
anthropic>=0.25.0
tavily-python>=0.3.3
httpx>=0.27.0
